Fixing new_tab script to work on new machine.
I don't know if it's Snow Leopard, new Visor, or some other variables, but the script
wasn't working as-is. The front window, instead of being the one where I'm working and
creating the new tab, is instead always the Visor window. Running new_tab gave me a new
tab in ~ and changed the directory in the Visor term. Not helpful.