Making a tiny .bashrc and pulling in nested files
Now the fun organization necessary setup is pretty much done. With a
tiny .bash_profile (that just pulls in .bashrc) and a tiny .bashrc (that
pulls in all the nested stuff + the special .local file), the rest is
just the actual organization work. And the old .bashrc is now pulled out
into its own setup file.