Start of cleanup script for iTunes DJ sets.
Those short silence tracks that I put between songs while DJing are very
handy, but getting rid of them when I'm done can be a pain. I want
something to take care of that for me. This is a start, but it errors
after a track is deleted from the playlist because then it tries to
access an index past the end of the shortened playlist.

Also, I was cute when I made the little silence tracks and named them
differing numbers of spaces (one space per second). Why is that?
Whatever the reason, I renamed them all to a single space for the name
check. It'd be better to make the check more specific and robust.

Next step: removing the divider ('----------') tracks, but only from the
end of the playlist. And hopefully (maybe?) only removing duplicates,
leaving one at the end (because sometimes I want that).