Add mvim_directory_server arguments for git
I like the convenience of 'v', but I dislike when I'm in some sort of
project that have to change directory to get a file to open in the right
instance. This lets me use 'vg' to use the git root as the instance
name.

If this becomes a problem outside of git repos, I'll figure out what
to do then.