Finder defaults
What do I want to show on the sidebar and the desktop.

Some of these values seem not entirely sticky, and for some (like the
preferred ordering of the favorites on the sidebar), it may be
prohibitively difficult to do it automatically rather than a bit of
manual dragging around. :hear_no_evil:

On the plus side, I don't need to logout/restart or even killall to get
new Finder windows showing the right target. :dizzy: