Fixing new_tab for ML
The `window 2` seemed to be an issue with a move to SL. (I'd have to dig
through the history of this script as it moved around, and I'm too tired
for that right now.)

Changing to `window 1` works, but it sometimes does a double `cd`, like
the cmd-T is doing it on its own. I will investigate further. And it may
require more changing after I install Visor.