add Atom keymap.cson file
This should probably be a symlink file handled entirely by the dotfiles,
but it needs to go into a certain directory. I'm considering what this
means for the Rakefile. Until then, I'll handle the placement using
boxen (which won't appear here). The important part is getting the file
under version control.