make Karabiner stop driving me crazy
I'm not sure why Karabiner uses a ~/.local *directory* to store its
stuff, and it appears that can't be changed. But I need to not see
complaints about this being a directory every time I open a new shell.

This should be better, looking for something in a different place(?) or
maybe accepting something in a directory, or whatever. But this'll do
for now.

Also, do I need to test -f *and* -r? I'm not sure. I'll figure it out
later.