add git ahead/behind to shell prompt
Just a little bit more info for my shell prompt.

Also making some sort of convenience, with a function to detect whether
I'm in a git repo. And putting a tiny bit of space in the prompt (before
the git-display, so the path and current-head don't run together.

I may need more space with the ahead/behind. We'll see.

Maybe time for colors soon, too.