make colors in git prompt display work
This has been a slog, mostly because of the way bash figures out how
long the prompt is (for wrapping and such). I can't have it all
separated out into different variables and functions and just include
itin PS1, but apparently PS1 needs the stuff right in there for the \[\]
sequence to show something is 0-length.

Welcome to the world of PROMPT_COMMAND setting PS1. (And so it needs to
share with what sets the window title.)

Going to follow this up with some cleanup about the unneeded function.

And this still seems to have a problem with the wrapping, in that it
happens at the right time/place (window/terminal edge), but it doesn't
go to the _next_ line. It just goes back to the beginning. It's a
carriage return, but not a line feed. Yay.