stomp prompt weirdness in fresh-init git repo
`git rev-parse HEAD` isn't happy in a repo without commits. One way to
handle this is change the prompt if the repo is fresh, just initted and
no commits yet. Another way – what I went with here – is to turn off
that part of the prompt in that situation. It's not *really* a repo yet,
after all. Is it?