video-encoding, leave file in place on error
Instead of saying it was completed, trying to do the rename, all that
noise, just complain and leave the file where it is.

I've been trying to figure out a good thing to do involving files with
spaces and also getting more options from the profile file. It's very
frustrating. I'll get there, and at least with my testing I don't want
to have to repeatedly rename files _back_ from the `complete` versions
to try again.