un-master my git alias assumptions
"main" is the new way to go, and there are a few things to consider
here. For one, I'm unlikely to go ahead and change the default/main
branch on every repo I own. For another, I can't really change this on
every repo I work on. With that in mind, be flexible but default to
'main' over 'master'. And figure out the correct value on each repo as
needed. (yay prompt commands)