adding some git rebase aliases
Now that I'm on the $MAIN_BRANCH kick, I might as well include an alias
or two to deal with that so I don't have to think about it as much.

And also some stuff to help me with the pain of dealing with conflicts
during a rebase.

And I guess let's consolidate all the rebase aliases to the same area.