make fetch-branch work even with the current branch
This involves a separate function that will check the current head and
compare it to the given branch. If they match, just pull instead of
trying the fetch trick.

This also involves having to run _bash_ as an interactive shell to get
the function that checks the current branch. And even then I'm using the
current-head one because it's the one that doesn't include the space at
the beginning. One day I'll have to get back to toying with awk and get
that output right.