fix git_current_branch leading space
There are apparently several things you can do to get rid of that
leading space, but I'm going with a classic *nix solution of adding
something more to the pipeline.

This also lets me actually check the current_branch output in
fetch_branch instead of looking at the current_head, which could be all
sorts of things that are sort of lies and helpful for the prompt.