handle PROMPT_COMMAND okay (prompted by tmux)
Okay, all the prompt stuff is going to happen in one place, and by that
I mean setting PS1 and also PROMPT_COMMAND. The functions may appear in
other places, but setting PROMPT_COMMAND is going to happen all
together.

Also, using a single function to add the command, to check if the
command needs to be added.

Starting up tmux was giving me some serious crap with the old way,
because it was just adding more instances of the same command to the
previous PROMPT_COMMAND value, which also included something that is
only provided by /etc/bashrc_Apple_Terminal, which I wasn't getting.