start off getting some defaults set
Making setting defaults part of the macos install, instead of a separate
part of updot.

Also starting to move over the defaults files from the osx dir. Doing it
piecemeal instead of all at once, and commenting out a bit that I'm not
sure about. Going to have to dig through this slowly.