get some media/webcam stuff set up
Very important here is finding out about uvcc, and the ability to set
things programmatically for my webcam that keeps needing its brightness
re-set.

The rest of this is just some organizing and dependencies, like uvcc is
node, so I need it and npm and to install it. I probably could do
something like a package.json and install uvcc that way. I'll look into
it.