get $DOTFILES_HOME working well
It doesn't need to (and shouldn't) end in a /. It was there for `find`
to work, but made for some confusing results (double-slash in the path
sometimes). It turns out all I needed was `find -H` and it was fine.

(going with -H because I don't think I want to follow symlinks in
general, but just with the command-line arguments)

Also getting a bit of consistency in the use of ${varname} vs. $varname,
and quoting.