update some git-prompt marker characters
I like the idea of the pilcrow for the pause. It's still like the
two-bar pause icon, but with some extra P fun.

Also, something better than * for dirty? And is reg mark too clever for
rebasing?

Don't _yet_ want to go full bash-git-prompt and breaking down how the
repo is dirty. But maybe I'll do that sometime.