get machine profiles in bootstrap and brew bundling
This is easy enough: in script/bootstrap, also look for a saved
~/.local/machine_profiles or prompt to populate it. In updot, make a
list of profiles with 'standard' plus the contents of that file, and
brew bundle for each one that has a corresponding Brewfile.

The pipeline for taking the user input and getting it to the file is
quite something, but it really works. I'm into it.

Also, I think 'profile' works better than 'type', especially with the
idea that there may be multiple.