add SSH config
This was all hidden away in a different place, but now I can definitely
have it all here together because of the Include directive.

Also I could be using `mkdir -p` for the ~/.ssh/ctrl stuff. I'll see if
I want to go in that direction. Supposedly it's important for ~/.ssh to
have particular permissions, but in practice (on macOS) it doesn't seem
to need that.