try linking symlink files on every updot
This might be pretty repetitive, but it's fine enough as far as I'm
concerned. It's not like running updot is something that's going to
happen all the time (like several times a day), _and_ it's not like
updot isn't already pretty repetitive, with the output from brew bundle
and all that.