try to make sure things show up in Keyboard preferences
This is a whole issue that I've seen elsewhere. There are places
`defaults write` doesn't work, because it's really writing to plist
files and those could not be writeable by the user. But sudo quiets that
error and yet does no actual writing? Maybe the file is still read-only
and even root can't write, but then wouldn't the error still show up?