add 'yarn' to work Brewfile
There's also something where apparently nvm is supposed to be installed
elsewhere, not via Homebrew. And therefore the Homebrew installation
doesn't put everything in the path. Currently I'm doing stuff in the
local bashrc because it's simple enough, but soon I may put something
directly in the node directory and have it conditionally loaded (if the
machine has a work profile).