allow setting a git repo's main branch
I'm starting to work on something where the "main branch" appears to be
"develop", and I don't want to add that yet as a standard choice. So
this seems to be a nice way to handle it — if you're getting the message
that a main branch can't be determined, just set it yourself.