add more media scripts
These are the things I used now, with Connect in the *arrs and running
as Custom Script.

For the most part, they work great. I'm not 100% happy about copying the
entire series when a single episode gets imported, but that's what I
got. And I do get single episodes at a time. And it's a `cp -al`, so
what's the problem?

Also finally did the auto-unmonitor, going through the effort of getting
python installed on my QNAP and using this API wrapper. It's not bad!

Do I really need to unmonitor as I go? No. Do I want to? Sure.